Enable bf16 and increase batch_size for A100 utilization
- Model now loads in bfloat16 (2x throughput on A100 tensor cores)
- batch_size: 2 → 8 (4x larger; better GRPO gradient estimates)
- max_epochs: 3 → 11 (6000÷8=750 steps/epoch, need 11 epochs to reach 8000 steps)
- max_steps=8000 is unchanged — still the actual stopping condition

+++ main.py
@@ -861,7 +861,8 @@ def get_model(config: DictConfig):
         model_cfg.name,
     )
     model = AutoModelForSeq2SeqLM.from_pretrained(
-        model_cfg.name
+        model_cfg.name,
+        torch_dtype=torch.bfloat16,
     )
 
     use_lora = bool(getattr(model_cfg, "use_lora", False))
